refactor: moved down filteredList.
