+++ src/containers/main/InvestContainer/index.tsx
@@ -1,14 +1,12 @@
-import React, { useCallback, useEffect, useState } from 'react';
+import React, { useCallback, useState } from 'react';
 import Box from '@mui/material/Box';
 import Tabs from '@mui/material/Tabs';
 import Tab from '@mui/material/Tab';
-import { useRouteMatch } from 'react-router-dom';
 import { UserSettings } from 'components/layout/UserSettings';
 import { InvestNav } from 'components/layout/InvestNav';
 import { useTranslation } from 'react-i18next';
-import { flowConfig, RoutesToFlowTypes } from 'constants/flowConfig';
 import { useShallowSelector } from 'hooks/useShallowSelector';
-import { selectMain, selectUserStreams } from 'store/main/selectors';
+import { selectMain } from 'store/main/selectors';
 import { RICAddress } from 'constants/polygon_config';
 import { useDispatch } from 'react-redux';
 import { startFlowAction, stopFlowAction } from 'store/main/actionCreators';
@@ -27,32 +25,10 @@ export const InvestContainer: React.FC<IProps> = () => {
 	const { t } = useTranslation();
 	const state = useShallowSelector(selectMain);
 	const { address, balances } = state;
-	const userStreams = useShallowSelector(selectUserStreams);
 	const dispatch = useDispatch();
-	const [filteredList, setFilteredList] = useState(flowConfig);
-	const match = useRouteMatch();
-	const routeEnd = match.path.slice(-7);
-	const flowType = RoutesToFlowTypes[match.path];
 
 	const [currentTab, setCurrentTab] = useState<TABS>(TABS.MARKET);
 	const switchTab = (evt: React.SyntheticEvent, tab: TABS) => setCurrentTab(tab);
-
-	useEffect(() => {
-		console.log(filteredList);
-	}, [filteredList]);
-
-	useEffect(() => {
-		if (flowType) {
-			setFilteredList(flowConfig.filter((each) => each.type === flowType));
-		} else {
-			const sortedUserStreams = userStreams.sort((a, b) => {
-				const flowA = parseFloat(state[a.flowKey]?.placeholder || '0');
-				const flowB = parseFloat(state[b.flowKey]?.placeholder || '0');
-				return flowB - flowA;
-			});
-			setFilteredList(sortedUserStreams);
-		}
-	}, [flowType, state, userStreams]);
 
 	const handleStart = useCallback(
 		(config: { [key: string]: string }) => (amount: string, callback: (e?: string) => void) => {
@@ -99,13 +75,7 @@ export const InvestContainer: React.FC<IProps> = () => {
 							</Tabs>
 						</Box>
 						<TabPanel index={TABS.MARKET} tab={currentTab}>
-							<InvestMarket
-								filteredList={filteredList}
-								setFilteredList={setFilteredList}
-								handleStart={handleStart}
-								handleStop={handleStop}
-								routeEnd={routeEnd}
-							/>
+							<InvestMarket handleStart={handleStart} handleStop={handleStop} />
 						</TabPanel>
 						<TabPanel index={TABS.INTERACTIVE} tab={currentTab}>
 							<InteractiveStreamManager />
